refactor(assistant): move auth-fallback-events-store out of memory/ into security/
