+++ assistant/src/__tests__/internal-telemetry-routes.test.ts
@@ -18,7 +18,6 @@ mock.module("../platform/consent-cache.js", () => ({
   getCachedShareAnalytics: () => shareAnalytics,
 }));
 
-import { queryUnreportedAuthFallbackEvents } from "../memory/auth-fallback-events-store.js";
 import { getDb } from "../persistence/db-connection.js";
 import { initializeDb } from "../persistence/db-init.js";
 import { authFallbackEvents } from "../persistence/schema/index.js";
@@ -26,6 +25,7 @@ import { GATEWAY_PRINCIPALS } from "../runtime/auth/route-policy.js";
 import { RouteError } from "../runtime/routes/errors.js";
 import { ROUTES } from "../runtime/routes/internal-telemetry-routes.js";
 import type { RouteHandlerArgs } from "../runtime/routes/types.js";
+import { queryUnreportedAuthFallbackEvents } from "../security/auth-fallback-events-store.js";
 
 await initializeDb();
