+++ assistant/src/security/auth-fallback-events-store.ts
